Move `grayscale` to `lib/thumbyGrayscale`
