+++ Games/GrayscaleTest/GrayscaleTest.py
@@ -13,11 +13,11 @@ from thumbyButton import actionPressed
 from machine import reset
 from time import sleep_ms
 from utime import ticks_us, sleep_us, ticks_diff
-import grayscale
+from thumbyGrayscale import Grayscale, Sprite
 
 # Initialization
 
-gs = grayscale.Grayscale()
+gs = Grayscale()
 
 # Helper function
 
@@ -59,7 +59,7 @@ waitKey()
 
 # Sprites as full screen images
 
-girlSprite = grayscale.Sprite(72, 40, bytearray([
+girlSprite = Sprite(72, 40, bytearray([
     128,4,160,8,130,32,8,162,0,40,130,8,160,10,64,39,53,187,234,149,106,181,214,253,135,1,81,44,210,40,74,180,192,40,129,1,135,207,191,254,232,218,144,96,1,131,7,13,30,59,237,246,217,38,219,171,94,173,106,225,148,64,8,130,32,8,162,0,40,2,144,4,
     0,84,0,18,64,136,2,168,2,80,10,64,10,160,5,168,1,85,135,30,189,122,127,103,37,134,17,0,133,64,3,1,11,141,6,181,20,31,15,87,79,215,255,255,216,32,193,160,112,208,248,243,239,159,124,235,212,187,69,186,215,94,248,200,130,32,10,160,5,80,4,0,
     0,149,32,133,80,8,66,40,2,169,4,81,8,162,8,146,36,129,40,130,104,213,74,136,224,224,192,68,198,201,134,192,137,64,2,224,16,232,182,125,255,255,255,255,107,206,63,191,222,127,71,241,215,127,244,0,3,15,61,126,233,246,219,255,16,74,0,74,32,5,168,1,
@@ -73,7 +73,7 @@ girlSprite = grayscale.Sprite(72, 40, bytearray([
     0,0,0,0,0,0,0,0,0,128,128,128,252,255,255,255,255,255,255,255,127,255,127,62,63,31,7,0,0,0,128,192,224,224,240,240,240,240,248,246,255,250,253,60,252,28,253,30,160,0,0,0,0,0,5,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0
 ]))
 
-parrotSprite = grayscale.Sprite(72, 40, bytearray([
+parrotSprite = Sprite(72, 40, bytearray([
     160,64,0,0,0,0,0,0,0,0,0,48,208,228,218,98,209,161,194,1,2,5,138,3,203,133,133,43,141,23,13,47,95,86,174,95,94,183,95,39,79,55,31,118,30,78,54,172,244,212,246,249,185,231,72,160,0,164,210,0,128,232,64,96,102,97,116,242,224,244,245,106,
     0,0,0,20,40,4,40,4,18,72,16,189,168,62,236,89,131,223,126,8,70,249,148,237,242,181,249,189,218,55,136,165,20,161,68,34,235,32,69,0,72,4,4,24,4,74,4,148,0,2,1,32,4,0,133,106,179,247,94,232,149,73,107,19,245,86,236,233,100,157,139,140,
     0,64,64,64,0,128,0,0,88,40,192,1,105,50,32,24,31,254,223,222,188,75,187,23,247,9,231,193,15,129,6,0,130,0,64,144,64,136,160,72,176,64,160,64,160,248,232,112,32,8,0,8,0,0,0,1,23,14,60,48,32,10,40,145,1,3,5,18,3,14,63,91,
@@ -97,7 +97,7 @@ waitKey()
 
 # Cat animation using a sprite
 
-cat = grayscale.Sprite(
+cat = Sprite(
     12, 9,         # Dimensions
     bytearray([    # Layer 2 data
         255,255,87,7,3,3,3,67,3,7,7,255,
