--- Games/GrayscaleTest/grayscale.py
@@ -1,1 +1,0 @@
-../../grayscale.py
