+++ Games/GrayscaleTest/thumbyGrayscale.py
@@ -1,0 +1,1 @@
+../../lib/thumbyGrayscale.py
