+++ lib/thumbyGrayscale.py
